Fasano-Franceschini Test

Performs a multidimensional two-sample Kolmogorov-Smirnov test as described by Fasano and Franceschini (1987). This test evaluates the null hypothesis that two i.i.d. random samples were drawn from the same underlying probability distribution. The data can be of any dimension, and can be of any type (continuous, discrete, or mixed).

Usage

fasano.franceschini.test(
  S1,
  S2,
  nPermute = 100,
  threads = 1,
  verbose = FALSE,
  seed = NULL,
  p.conf.level = 0.95
)

Arguments

S1

matrix or data.frame.

S2

matrix or data.frame.

nPermute

a nonnegative integer setting the number of permuted samples to generate when estimating the permutation test p-value. Default is 100. If set to 0, no p-value is estimated.

threads

a positive integer or "auto" setting the number of threads used for performing the permutation test. If set to "auto", the number of threads is determined by RcppParallel::defaultNumThreads(). Default is 1.

verbose

a boolean indicating whether to display a progress bar. Default is FALSE. Only available for serial version (threads = 1).

seed

optional integer to seed the PRNG used for the permutation test. Default is NULL. Only available for serial version (threads = 1).

p.conf.level

confidence level for the confidence interval of the permutation test p-value.

Value

A list with class htest containing the following components:

statistic

the value of the test statistic Z.

estimate

the value of the sample test statistics D1 and D2.

p.value

the permutation test p-value.

conf.int

a binomial confidence interval for the p-value.

method

a character string indicating what type of test was performed.

data.name

a character string giving the names of the data.

Details

The test statistic is computed using an efficient range tree implementation with a time complexity of O(n*log(n)^(d-1)), where n is the size of the largest sample and d is the dimension of the data.

The p-value for the test is computed empirically using a permutation test. As it is almost always infeasible to compute the exact permutation test p-value, a Monte Carlo approximation is made instead. This estimate is a binomially distributed random variable, and thus a confidence interval can be computed. The confidence interval is obtained using the procedure given in Clopper and Pearson (1934).

Since p-values are calculated using a permutation test, the data can be of any dimensionality and of any type (continuous, discrete, or mixed). If there are ties in the data, points are counted with multiplicity.

References

Fasano, G. & Franceschini, A. (1987). A multidimensional version of the Kolmogorov-Smirnov test. Monthly Notices of the Royal Astronomical Society, 225:155-170. doi:10.1093/mnras/225.1.155 .

Clopper, C. J. & Pearson, E. S. (1934). The use of confidence or fiducial limits illustrated in the case of the binomial. Biometrika, 26, 404–413. doi:10.2307/2331986 .

Examples

set.seed(0)

# create 2-D samples using data frames
S1 <- data.frame(x = rnorm(n = 50, mean = 1, sd = 2),
                 y = rnorm(n = 50, mean = 3, sd = 1))
S2 <- data.frame(x = rnorm(n = 150, mean = 1, sd = 2),
                 y = rnorm(n = 150, mean = 3, sd = 1))

# perform test (serial version)
fasano.franceschini.test(S1, S2)
#> 
#> 	Fasano-Francheschini Test
#> 
#> data:  S1 and S2
#> Z = 0.87773, p-value = 0.8515
#> 95 percent confidence interval:
#>  0.7669031 0.9144277
#> sample estimates:
#>        D1        D2 
#> 0.1333333 0.1533333 
#> 
# perform test with more permutations
fasano.franceschini.test(S1, S2, nPermute = 200)
#> 
#> 	Fasano-Francheschini Test
#> 
#> data:  S1 and S2
#> Z = 0.87773, p-value = 0.8159
#> 95 percent confidence interval:
#>  0.7553133 0.8669551
#> sample estimates:
#>        D1        D2 
#> 0.1333333 0.1533333 
#> 
# set seed for reproducible p-value
fasano.franceschini.test(S1, S2, seed = 0)
#> 
#> 	Fasano-Francheschini Test
#> 
#> data:  S1 and S2
#> Z = 0.87773, p-value = 0.8911
#> 95 percent confidence interval:
#>  0.8134799 0.9443628
#> sample estimates:
#>        D1        D2 
#> 0.1333333 0.1533333 
#> 
# display progress bar
fasano.franceschini.test(S1, S2, verbose = TRUE)
#> 
#> 	Fasano-Francheschini Test
#> 
#> data:  S1 and S2
#> Z = 0.87773, p-value = 0.8614
#> 95 percent confidence interval:
#>  0.7783773 0.9220964
#> sample estimates:
#>        D1        D2 
#> 0.1333333 0.1533333 
#> 
# change confidence level for p-value confidence interval
fasano.franceschini.test(S1, S2, p.conf.level = 0.99)
#> 
#> 	Fasano-Francheschini Test
#> 
#> data:  S1 and S2
#> Z = 0.87773, p-value = 0.8416
#> 99 percent confidence interval:
#>  0.7274736 0.9221261
#> sample estimates:
#>        D1        D2 
#> 0.1333333 0.1533333 
#> 

# perform test (parallel version, 2 threads)
if (FALSE) {
fasano.franceschini.test(S1, S2, threads = 2)}

# create 3-D mixed samples using matrices
S1 <- cbind(rgamma(n = 43, shape = 2),
            rpois(n = 43, lambda = 5),
            rpois(n = 43, lambda = 3.5))
S2 <- cbind(rgamma(n = 72, shape = 2),
            rpois(n = 72, lambda = 5),
            rpois(n = 72, lambda = 5))

# perform test
fasano.franceschini.test(S1, S2)
#> 
#> 	Fasano-Francheschini Test
#> 
#> data:  S1 and S2
#> Z = 1.8443, p-value = 0.009901
#> 95 percent confidence interval:
#>  0.00025064 0.05393235
#> sample estimates:
#>        D1        D2 
#> 0.3323643 0.3785530 
#>