Fasano Franceschini Test

Computes the 2-D Kolmogorov-Smirnov two-sample test as described by Fasano and Franceschini (1987).

fasano.franceschini.test(S1, S2, nBootstrap = 0, cores = 2)

Arguments

S1
a n by 2 data.frame of x and y coordinates of sample 1

S2
a n by 2 data.frame of x and y coordinates of sample 2

nBootstrap
a numeric defining the number of bootstrapped samples to be generated for computing the empirical p-value (note this procedure is slow and computationally expensive on the order of nBootStrap*O(n^2). Default is set to 0. If nBootstrap is 0, the Fasano Franceschini distributional approximation is used for defining the p-value. See Fasano and Franceschini test (1987) for details.

cores
a numeric defining the number of cores to use of processing

Value

the 2-D ks statistic and p-value

Details

Code adapted from Press, W. H., Teukolsky, S. A., Vetterling, W. T.,, Flannery, B. P. (2007). Numerical Recipes 3rd Edition: The Art of Scientific Computing. Cambridge University Press. ISBN: 0521880688

References

Fasano, G., Franceschini, A. (1987)DOI:10.1093/mnras/225.1.155. A multidimensional version of the Kolmogorov-Smirnov test. Monthly Notices of the Royal Astronomical Society 225:155-170.

Peacock J.A. (1983) DOI:10.1093/mnras/202.3.615. Two-dimensional goodness-of-fit testing in astronomy. Monthly Notices of the Royal Astronomical Society 202:615-627.

Press, W. H., Teukolsky, S. A., Vetterling, W. T.,, Flannery, B. P. (2007). Numerical Recipes 3rd Edition: The Art of Scientific Computing. Cambridge University Press. ISBN: 0521880688